requirements.txt
discord.py
python-dotenv
openpyxl